IMPORTANT: All these features aren't tested yet and could be removed or changed at any moment. =========
Proof of concept:
----------------
Create the ExtDirectCRUD base class in order to implement the basic CRUD actions and use it with an Ext.data.DirectStore in the client-side.

The example that we want to cover it's available in: http://www.extjs.com/deploy/dev/examples/writer/writer.html

The simplest usage it's:

@crud(remote_provider)
class PeopleCRUD(ExtDirectCRUD):
    model = PeopleModel

and that's all you need to write for the server-side (django). The ExtJS code, it's the same from the example with minor changes needed to use Ext.Direct.

Also, we added a few optional parameters to the ExtDirectStore class in order to generate the `metaData` property and configure the JsonReader in
the client-side.